requirement.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.2.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
html5lib==1.1
idna==3.4
lxml==4.9.2
outcome==1.2.0
PyPDF2==3.0.1
PySocks==1.7.1
pytz==2023.3
redis==4.5.5
requests==2.30.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.6
webencodings==0.5.1
wsproto==1.2.0
